+++ src/cartridges/int_adyen_SFRA/cartridge/adyen/webhooks/libs/libAuthenticationUtils.js
@@ -24,7 +24,7 @@ const Encoding = require('dw/crypto/Encoding');
 const Mac = require('dw/crypto/Mac');
 const Bytes = require('dw/util/Bytes');
 const constants = require('*/cartridge/adyenConstants/constants');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 const AdyenConfigs = require('*/cartridge/scripts/util/adyenConfigs');
 
 /**
