+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/adyen/__tests__/notify.test.js
@@ -18,19 +18,19 @@ afterEach(() => {
 
 describe('Notify', () => {
   it('should render error when status is falsy', () => {
-    const checkAuth = require('*/cartridge/scripts/checkNotificationAuth');
+    const checkAuth = require('*/cartridge/adyen/webhooks/checkNotificationAuth');
     checkAuth.check.mockImplementation(() => false);
     notify(req, res, jest.fn());
     expect(res.render).toHaveBeenCalledWith('/adyen/error');
   });
   it('should render notify when notification result is successful', () => {
-    const handleNotify = require('*/cartridge/scripts/handleNotify');
+    const handleNotify = require('*/cartridge/adyen/webhooks/handleNotify');
     handleNotify.notify.mockImplementation(() => ({ success: true }));
     notify(req, res, jest.fn());
     expect(res.render).toHaveBeenCalledWith('/notify');
   });
   it('should render notifyError when notification result is not successful', () => {
-    const handleNotify = require('*/cartridge/scripts/handleNotify');
+    const handleNotify = require('*/cartridge/adyen/webhooks/handleNotify');
     handleNotify.notify.mockImplementation(() => ({
       success: false,
       errorMessage: 'mocked_error_message',
