+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/adyen/checkBalance.js
@@ -6,7 +6,7 @@ const AdyenHelper = require('*/cartridge/scripts/util/adyenHelper');
 const adyenCheckout = require('*/cartridge/scripts/adyenCheckout');
 const AdyenConfigs = require('*/cartridge/scripts/util/adyenConfigs');
 const constants = require('*/cartridge/adyenConstants/constants');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 
 function getFormattedProperties(checkBalanceResponse, orderAmount) {
   if (checkBalanceResponse.resultCode === 'Success') {
