+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/adyen/index.js
@@ -1,6 +1,6 @@
 const showConfirmation = require('*/cartridge/controllers/middlewares/adyen/showConfirmation');
 const paymentFromComponent = require('*/cartridge/controllers/middlewares/adyen/paymentFromComponent');
-const notify = require('*/cartridge/controllers/middlewares/adyen/notify');
+const notify = require('*/cartridge/adyen/webhooks/notify');
 const showConfirmationPaymentFromComponent = require('*/cartridge/controllers/middlewares/adyen/showConfirmationPaymentFromComponent');
 const paymentsDetails = require('*/cartridge/controllers/middlewares/adyen/paymentsDetails');
 const redirect3ds1Response = require('*/cartridge/controllers/middlewares/adyen/redirect3ds1Response');
