+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/adyen/partialPayment.js
@@ -5,7 +5,7 @@ const Resource = require('dw/web/Resource');
 const AdyenConfigs = require('*/cartridge/scripts/util/adyenConfigs');
 const adyenCheckout = require('*/cartridge/scripts/adyenCheckout');
 const AdyenHelper = require('*/cartridge/scripts/util/adyenHelper');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 const constants = require('*/cartridge/adyenConstants/constants');
 
 function responseContainsErrors(response) {
