+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/adyen/paymentsDetails.js
@@ -3,7 +3,7 @@ const OrderMgr = require('dw/order/OrderMgr');
 const Transaction = require('dw/system/Transaction');
 const AdyenHelper = require('*/cartridge/scripts/util/adyenHelper');
 const adyenCheckout = require('*/cartridge/scripts/adyenCheckout');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 
 function getSignature(paymentsDetailsResponse, orderToken) {
   const order = OrderMgr.getOrder(
