+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/adyen/redirect3ds1Response.js
@@ -1,7 +1,7 @@
 const URLUtils = require('dw/web/URLUtils');
 const adyenCheckout = require('*/cartridge/scripts/adyenCheckout');
 const constants = require('*/cartridge/adyenConstants/constants');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 
 /*
  * Redirects to list of added cards on success. Otherwise redirects to add payment with error
