+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/adyen/saveExpressShopperDetails.js
@@ -1,7 +1,7 @@
 const URLUtils = require('dw/web/URLUtils');
 const Transaction = require('dw/system/Transaction');
 const BasketMgr = require('dw/order/BasketMgr');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 const AdyenHelper = require('*/cartridge/scripts/util/adyenHelper');
 
 function setBillingAndShippingAddress(currentBasket) {
