+++ src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/checkout/begin.js
@@ -6,7 +6,7 @@ const URLUtils = require('dw/web/URLUtils');
 const AdyenConfigs = require('*/cartridge/scripts/util/adyenConfigs');
 const AdyenHelper = require('*/cartridge/scripts/util/adyenHelper');
 const { updateSavedCards } = require('*/cartridge/scripts/updateSavedCards');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 
 function shouldRestoreBasket(cachedOrderNumber) {
   // restore cart if order number was cached
