+++ src/cartridges/int_adyen_SFRA/cartridge/scripts/adyenDeleteRecurringPayment.js
@@ -23,7 +23,7 @@
 const AdyenHelper = require('*/cartridge/scripts/util/adyenHelper');
 const AdyenConfigs = require('*/cartridge/scripts/util/adyenConfigs');
 const constants = require('*/cartridge/adyenConstants/constants');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 
 function deleteRecurringPayment(args) {
   try {
