+++ src/cartridges/int_adyen_SFRA/cartridge/scripts/adyenGetPaymentMethods.js
@@ -24,7 +24,7 @@ const AdyenHelper = require('*/cartridge/scripts/util/adyenHelper');
 const AdyenConfigs = require('*/cartridge/scripts/util/adyenConfigs');
 const constants = require('*/cartridge/adyenConstants/constants');
 const blockedPayments = require('*/cartridge/scripts/config/blockedPaymentMethods.json');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 
 function getMethods(basket, customer, countryCode) {
   try {
