+++ src/cartridges/int_adyen_SFRA/cartridge/scripts/adyenGiving.js
@@ -25,7 +25,7 @@ const Transaction = require('dw/system/Transaction');
 const AdyenHelper = require('*/cartridge/scripts/util/adyenHelper');
 const AdyenConfigs = require('*/cartridge/scripts/util/adyenConfigs');
 const constants = require('*/cartridge/adyenConstants/constants');
-const AdyenLogs = require('*/cartridge/scripts/adyenCustomLogs');
+const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 
 function donate(donationReference, donationAmount, orderToken) {
   try {
